ref(back): restructure implementing routes
