+++ backend/src/router/userRouter.py
@@ -1,0 +1,107 @@
+from http import HTTPStatus
+
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy import select
+from sqlalchemy.exc import IntegrityError
+from sqlalchemy.orm import Session
+
+from src.database import get_session
+from src.models import User
+from src.schemas import Message, UserList, UserPublic, UserSchema
+from src.security import (
+    get_current_user,
+    get_password_hash,
+)
+
+router = APIRouter(prefix='/user', tags=['user'])
+
+
+@router.get('/', response_model=UserList)
+def read_users(
+    skip: int = 0, limit: int = 100, session: Session = Depends(get_session)
+):
+    users = session.scalars(select(User).offset(skip).limit(limit)).all()
+    return {'users': users}
+
+
+@router.post('/', status_code=HTTPStatus.CREATED, response_model=UserPublic)
+def create_user(user: UserSchema, session: Session = Depends(get_session)):
+    db_user = session.scalar(
+        select(User).where(
+            (User.user_nickname == user.user_nickname)
+            | (User.user_email == user.user_email)
+        )
+    )
+
+    if db_user:
+        if db_user.user_nickname == user.user_nickname:
+            raise HTTPException(
+                status_code=HTTPStatus.CONFLICT,
+                detail='Nickname already exists',
+            )
+        elif db_user.user_email == user.user_email:
+            raise HTTPException(
+                status_code=HTTPStatus.CONFLICT,
+                detail='Email already exists',
+            )
+
+    hashed_password = get_password_hash(user.user_password)
+
+    db_user = User(
+        user_name=user.user_name,
+        user_nickname=user.user_nickname,
+        user_email=user.user_email,
+        user_password=hashed_password,
+    )
+    session.add(db_user)
+    session.commit()
+    session.refresh(db_user)
+
+    return db_user
+
+
+@router.put('/{user_id}', response_model=UserPublic)
+def update_user(
+    user_id: int,
+    user: UserSchema,
+    session: Session = Depends(get_session),
+    current_user: User = Depends(get_current_user),
+):
+    if current_user.user_id != user_id:
+        raise HTTPException(
+            status_code=HTTPStatus.FORBIDDEN, detail='Not enough permissions'
+        )
+
+    try:
+        current_user.user_name = user.user_name
+        current_user.user_nickname = user.user_nickname
+        current_user.user_password = get_password_hash(user.user_password)
+        current_user.user_email = user.user_email
+        session.commit()
+        session.refresh(current_user)
+
+        return current_user
+
+    except IntegrityError:
+        session.rollback()
+        raise HTTPException(
+            status_code=HTTPStatus.CONFLICT,
+            detail='Nickname or Email already exists',
+        )
+
+
+@router.delete('/{user_id}', response_model=Message)
+def delete_user(
+    user_id: int,
+    session: Session = Depends(get_session),
+    current_user: User = Depends(get_current_user),
+):
+    if current_user.user_id != user_id:
+        raise HTTPException(
+            status_code=HTTPStatus.FORBIDDEN, detail='Not enough permissions'
+        )
+
+    session.delete(current_user)
+    session.commit()
+
+    return {'message': 'User deleted'}
