+++ backend/tests/conftest.py
@@ -79,7 +79,7 @@ def user(session):
 @pytest.fixture
 def token(client, user):
     response = client.post(
-        '/token',
+        '/auth/token',
         data={
             'username': user.user_email,
             'password': user.clear_password,
